Move summary key type tests to corresponding file
